# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Conduit API",
        "description": "Conduit is a social blogging site (like Medium) built on top of Laravel. This is the API backend.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "Conduit API server"
        }
    ],
    "paths": {
        "/articles": {
            "get": {
                "tags": [
                    "Articles"
                ],
                "summary": "Get recent articles globally.",
                "description": "Get most recent articles globally. Use query parameters to filter results. Auth is optional",
                "operationId": "GetArticles",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MultipleArticlesResponse"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Articles"
                ],
                "summary": "Create an article.",
                "description": "Create an article. Auth is required",
                "operationId": "CreateArticle",
                "requestBody": {
                    "description": "Article to create",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/NewArticleRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SingleArticleResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "scalar": "BearerToken"
                    }
                ]
            }
        },
        "/articles/feed": {
            "get": {
                "tags": [
                    "Articles"
                ],
                "summary": "Get recent articles from users you follow.",
                "description": "Get most recent articles from users you follow. Use query parameters to limit. Auth is required",
                "operationId": "GetArticlesFeed",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MultipleArticlesResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "scalar": "BearerToken"
                    }
                ]
            }
        },
        "/articles/{slug}": {
            "get": {
                "tags": [
                    "Articles"
                ],
                "summary": "Get an article.",
                "description": "Get an article. Auth not required",
                "operationId": "GetArticle",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "description": "Slug of the article to get",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SingleArticleResponse"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Articles"
                ],
                "summary": "Update an article.",
                "description": "Update an article. Auth is required",
                "operationId": "UpdateArticle",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "description": "Slug of the article to update",
                        "required": true
                    }
                ],
                "requestBody": {
                    "description": "Article to update",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateArticleRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SingleArticleResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "scalar": "BearerToken"
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Articles"
                ],
                "summary": "Delete an article.",
                "description": "Delete an article. Auth is required",
                "operationId": "DeleteArticle",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "description": "Slug of the article to delete",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Success"
                    }
                },
                "security": [
                    {
                        "scalar": "BearerToken"
                    }
                ]
            }
        },
        "/articles/{slug}/favorite": {
            "post": {
                "tags": [
                    "Favorites"
                ],
                "summary": "Favorite an article.",
                "description": "Favorite an article. Auth is required",
                "operationId": "CreateArticleFavorite",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "description": "Slug of the article that you want to favorite",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SingleArticleResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "scalar": "BearerToken"
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Favorites"
                ],
                "summary": "Unfavorite an article.",
                "description": "Unfavorite an article. Auth is required",
                "operationId": "DeleteArticleFavorite",
                "parameters": [
                    {
                        "name": "slug",
                        "in": "path",
                        "description": "Slug of the article that you want to unfavorite",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SingleArticleResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "scalar": "BearerToken"
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "NewArticle": {
                "properties": {
                    "title": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "body": {
                        "type": "string"
                    },
                    "tagList": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    }
                },
                "type": "object"
            },
            "NewArticleRequest": {
                "properties": {
                    "article": {
                        "$ref": "#/components/schemas/NewArticle"
                    }
                },
                "type": "object"
            },
...
            "UpdateArticle": {
                "properties": {
                    "title": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "body": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "UpdateArticleRequest": {
                "properties": {
                    "article": {
                        "$ref": "#/components/schemas/UpdateArticle"
                    }
                },
                "type": "object"
            },
...
            "MultipleArticlesResponse": {
                "properties": {
                    "articles": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Article"
                        }
                    },
                    "articlesCount": {
                        "type": "integer"
                    }
                },
                "type": "object"
            },
...
            "Profile": {
                "properties": {
                    "username": {
                        "type": "string"
                    },
                    "bio": {
                        "type": "string"
                    },
                    "image": {
                        "type": "string"
                    },
                    "following": {
                        "type": "boolean",
                        "default": false
                    }
                },
                "type": "object"
            },
...
            "Article": {
                "properties": {
                    "title": {
                        "type": "string"
                    },
                    "slug": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "body": {
                        "type": "string"
                    },
                    "createdAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updatedAt": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "author": {
                        "$ref": "#/components/schemas/Profile"
                    },
                    "tagList": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "favorited": {
                        "type": "boolean",
                        "default": false
                    },
                    "favoritesCount": {
                        "type": "integer"
                    }
                },
                "type": "object"
            },
            "SingleArticleResponse": {
                "properties": {
                    "article": {
                        "$ref": "#/components/schemas/Article"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "BearerToken": {
                "type": "apiKey",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}